backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-multipart>=0.0.9
